requirements.txt
torch
torchvision
timm
streamlit
pillow